Change the way density is compute in weakly-compressible boussinesq problem
Before, the density dependance on temperature was only taken into account in the \rho\vec{b} term (like in the incompressible term). So that the "density" was the pressure part of the density. To be more accurate the density now dpends on the temperature directly (rho = f(p)*g(T))